+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## next
 
-- Add the ability to view hook events for a given subscription `hook events -s [subscription_id]`.
+## 0.0.18
+
 - Add the ability to list participant groups via `participant list -p [project_id]`.
 - Add the ability to view a participant group via `participant view [group_id]`.
+
+## 0.0.17
+
+- Add the ability to view hook events for a given subscription `hook events -s [subscription_id]`.
 
 ## 0.0.16